requirements.txt
streamlit
pandas
openpyxl
python-docx
pdfplumber
python-dotenv
google-genai
google-adk
opentelemetry-instrumentation-google-genai
sqlalchemy